Moved Breed into separate Table as Breed is Entity
