+++ FamilyPets.Backend/src/FamilyPets.Domain/VolunteerAgregate/PetValueObjects/PetBreedAndSpecies.cs
@@ -1,5 +1,5 @@
 ﻿using CSharpFunctionalExtensions;
-using FamilyForPets.Domain.Species;
+using FamilyForPets.Domain.SpeciesAgregate;
 
 namespace FamilyForPets.Domain.VolunteerAgregate.PetValueObjects
 {
