+++ FamilyPets.Backend/src/FamilyPets.Infrastructure/Configurations/BreedConfiguration.cs
@@ -10,28 +10,23 @@ using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
 namespace FamilyForPets.Infrastructure.Configurations
 {
-    public class SpeciesConfiguration : IEntityTypeConfiguration<Species>
+    public class BreedConfiguration : IEntityTypeConfiguration<Breed>
     {
-        public void Configure(EntityTypeBuilder<Species> builder)
+        public void Configure(EntityTypeBuilder<Breed> builder)
         {
-            builder.ToTable("species");
+            builder.ToTable("breeds");
 
             builder.HasKey(s => s.Id);
 
             builder.Property(s => s.Id)
                 .HasConversion(
                     id => id.Value,
-                    value => SpeciesId.Create(value));
+                    value => BreedId.Create(value));
 
             builder.Property(s => s.Name)
                 .IsRequired()
                 .HasMaxLength(Species.MAX_NAME_LENGHT);
-
-            builder.HasMany(v => v.Breeds)
-                .WithOne()
-                .HasForeignKey("species_id")
-                .OnDelete(DeleteBehavior.Cascade)
-                .IsRequired();
         }
     }
+
 }
