+++ FamilyPets.Backend/src/FamilyPets.Infrastructure/Configurations/Converters/ColorToInt32Converter.cs
@@ -1,11 +1,7 @@
 ﻿using System.Drawing;
-using FamilyForPets.Domain.VolunteerAgregate;
-using FamilyForPets.Domain.VolunteerAgregate.PetValueObjects;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata.Builders;
 using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 
-namespace FamilyForPets.Infrastructure.Configurations
+namespace FamilyForPets.Infrastructure.Configurations.Converters
 {
     public class ColorToInt32Converter : ValueConverter<Color, int>
     {
